backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
pydantic
pydantic-settings
yfinance==0.2.40

python-dotenv
httpx
alembic
asyncpg
greenlet
aiosqlite
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
bcrypt==4.0.1
